+++ src/main/BoxStyleConstants.ts
@@ -69,7 +69,7 @@ const INNER_THIN_OUTER_THICK = new BoxStyle(
 const INNER_THICK_OUTER_THIN = new BoxStyle(
   ['\u2501', '\u2500', '\u2503', '\u2502'],
   ['\u250c', '\u2510', '\u2514', '\u2518'],
-  '\u2548',
+  '\u254B',
   ['\u2530', '\u2538', '\u251d', '\u2525']
 );
 const INNER_DOUBLE_OUTER_THIN_CURVE = new BoxStyle(
@@ -81,7 +81,7 @@ const INNER_DOUBLE_OUTER_THIN_CURVE = new BoxStyle(
 const INNER_THICK_OUTER_THIN_CURVE = new BoxStyle(
   ['\u2501', '\u2500', '\u2503', '\u2502'],
   ['\u256d', '\u256e', '\u2570', '\u256f'],
-  '\u2548',
+  '\u254B',
   ['\u2530', '\u2538', '\u251d', '\u2525']
 );