+++ CHANGELOG.md
@@ -3,6 +3,8 @@
 ## `0.0.8`
 - refactor Simple Tests
 - fix null or '' error
+- refactor Array Tests
+- update README.md
 
 ## `0.0.7` 28 Dec 2022
 - add Horizontal/Vertical inner and outer